requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
python-engineio==4.8.0
eventlet==0.33.3
python-dotenv==1.0.0
